Error handling

+++ app/Services/RobotService.php
@@ -2,16 +2,23 @@
 
 namespace App\Services;
 
-use App\Contracts\RepositoryInterface;
+
 use App\Robot;
 use Carbon\Carbon;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
+use App\Contracts\RepositoryInterface;
+use Illuminate\Support\Facades\Storage;
+use \Illuminate\Database\QueryException;
+use App\Exceptions\RobotService\RobotNotFoundException;
+use App\Exceptions\RobotService\RobotBulkStructureException;
+use App\Exceptions\RobotService\RobotOwnerMismatchedException;
+use App\Exceptions\RobotService\RobotBulkDataErrorException;
+
 
 class RobotService implements RepositoryInterface
 {
     /**
-     * Get all robots.
+     * Get all Robots.
      *
      * @return \Illuminate\Database\Eloquent\Collection
      */
@@ -29,11 +36,12 @@ class RobotService implements RepositoryInterface
      */
     public function create($request) : Robot
     {
+        $request['created_by'] = $request['updated_by'] = $request['user_id'] = Auth::user()->id;
         return Robot::create($request); 
     }
 
     /**
-     * Delete a robot by id.
+     * Delete a Robot by id.
      *
      * @param $id
      *
@@ -41,11 +49,21 @@ class RobotService implements RepositoryInterface
      */
     public function delete($id)
     {
+        $robot = Robot::find($id);
+        if(!$robot)
+        {
+            throw new RobotNotFoundException();
+        }
+        if($robot->user_id != Auth::id())
+        {
+            throw new RobotOwnerMismatchedException();
+        }
+
         return Robot::destroy($id);
     }
 
     /**
-     * Update a robot.
+     * Update a Robot.
      *
      * @param $request
      * @param $id
@@ -54,13 +72,21 @@ class RobotService implements RepositoryInterface
      */
     public function update($request, $id) 
     {
-        //return Robot::where('id', $id)->update($request);
+        $robot = Robot::find($id);
+        if(!$robot)
+        {
+            throw new RobotNotFoundException();
+        }
+        if($robot->user_id != Auth::id())
+        {
+            throw new RobotOwnerMismatchedException();
+        }
 
         return tap(Robot::findOrFail($id))->update($request)->fresh();
     }
 
     /**
-     * Find a robot by id.
+     * Find a Robot by id.
      *
      * @param $id
      *
@@ -68,7 +94,12 @@ class RobotService implements RepositoryInterface
      */
     public function find($id)
     {
-        return Robot::find($id);
+        $robot = Robot::find($id);
+        if(!$robot)
+        {
+            throw new RobotNotFoundException();
+        }
+        return $robot;
     }
 
     /**
@@ -80,27 +111,38 @@ class RobotService implements RepositoryInterface
      */
     public function createBulk($request)
     {
-        if($request->hasFile('file'))
-        {
-            $file = $request->file('file');
-            Storage::disk('local')->putFileAs('', $file, $file->getClientOriginalName());
+        $requiredStructure = ['name', 'power', 'speed', 'weight'];
+        $file = $request->file('file');
+        Storage::disk('local')->putFileAs('', $file, $file->getClientOriginalName());
+        
+        $lines = explode("\n", file_get_contents($file));
+        $head = str_getcsv(array_shift($lines));
+        sort($head);
 
-            
-            $lines = explode("\r\n", file_get_contents($request->file('file')));
-            $head = str_getcsv(array_shift($lines));
-            $robots = [];
+        $head !== $requiredStructure ? throw new RobotBulkStructureException() : 5;
+        // if($head !== $requiredStructure)
+        // {
+        //     throw new RobotBulkStructureException();
+        // }
 
-            for ($i = 0; $i < count($lines); $i++) 
-            { 
-                if($i == 0 || !strlen($lines[$i])) continue;
-                $check = array_combine($head, str_getcsv($lines[$i]));
-                $check['created_by'] = Auth::user()->id;
-                $check['updated_by'] = Auth::user()->id;
-                $check['created_at'] = now();
-                $check['updated_at'] = now();
-                $robots[] = $check;
-            }
-            return Robot::insert($robots);
+        $robots = [];
+        for ($i = 0; $i < count($lines); $i++) 
+        { 
+            if($i == 0 || !strlen($lines[$i])) continue;
+            $check = array_combine($head, str_getcsv($lines[$i]));
+            $check['created_by'] = $check['updated_by'] = $check['user_id'] = Auth::id();
+            $check['created_at'] = $check['updated_at'] = now();
+            $robots[] = $check;
         }
+        
+        try 
+        {
+            Robot::insert($robots);
+        }
+        catch(QueryException $exception)
+        {
+            throw new RobotBulkDataErrorException();
+        }
+        return true;
     }
 }
